+++ client/src/components/drawerviews/CardDrawerView.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function ColumnDrawerView(props) {
+  return (
+    <Fade right when={context[0].open} collapse duration={400}>
+      <div>Card Edit</div>
+    </Fade>
+  );
+}
