+++ client/src/components/drawerviews/ColumnDrawerView.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function ColumnDrawerView(props) {
+  return (
+    <Fade right when={context[0].open} collapse duration={400}>
+      <div>Column Edit</div>
+    </Fade>
+  );
+}
